Initial BinaryInstaller extracted from Project.swift
