+++ Tests/CarthageKitTests/BinaryInstallerSpec.swift
@@ -1,0 +1,8 @@
+@testable import CarthageKit
+import Foundation
+import Quick
+import Nimble
+
+final class BinaryInstallerSpec: QuickSpec {
+    
+}
